requirements.txt

discord.py>=2.3.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
psutil>=5.9.0
